+++ model/otlpgrpc/metrics.go
@@ -20,9 +20,9 @@ import (
 
 	"google.golang.org/grpc"
 
-	"go.opentelemetry.io/collector/model/internal"
 	otlpcollectormetrics "go.opentelemetry.io/collector/model/internal/data/protogen/collector/metrics/v1"
 	otlpmetrics "go.opentelemetry.io/collector/model/internal/data/protogen/metrics/v1"
+	ipdata "go.opentelemetry.io/collector/model/internal/pdata"
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
@@ -111,11 +111,11 @@ func (mr MetricsRequest) MarshalJSON() ([]byte, error) {
 }
 
 func (mr MetricsRequest) SetMetrics(ld pdata.Metrics) {
-	mr.orig.ResourceMetrics = internal.MetricsToOtlp(ld.InternalRep()).ResourceMetrics
+	mr.orig.ResourceMetrics = ipdata.MetricsToOtlp(ld).ResourceMetrics
 }
 
 func (mr MetricsRequest) Metrics() pdata.Metrics {
-	return pdata.MetricsFromInternalRep(internal.MetricsFromOtlp(&otlpmetrics.MetricsData{ResourceMetrics: mr.orig.ResourceMetrics}))
+	return ipdata.MetricsFromOtlp(&otlpmetrics.MetricsData{ResourceMetrics: mr.orig.ResourceMetrics})
 }
 
 // MetricsClient is the client API for OTLP-GRPC Metrics service.