+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🛑 Breaking changes 🛑
 
 - Remove `Type` funcs in pdata (#4933)
+- pdata: deprecate funcs working with InternalRep (#4957)
 
 ## 🧰 Bug fixes 🧰